feat: перенос конфигурации Next.js на JS
