+++ BUILD_REPORT.md
@@ -107,3 +107,8 @@
 - Workflow CI уточнён: реальные переменные Yandex Registry, шаг деплоя помечен TODO.
 - Добавлена переменная USE_MOCK_DATA и примеры мок-данных для сборки без API.
 - Создан общий `.env.example`, расширен `.gitignore`.
+
+## 2025-06-30
+
+- Перенесён next.config на JavaScript, удалены все упоминания `.ts`.
+- Все тесты пройдены после миграции.
