+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 - Создан файл `CONTRIBUTING.md`.
 - Обновлён CI workflow и документация.
 - Добавлен `.env.example`, улучшен `.gitignore`, API функции поддерживают мок-данные.
+
+## 2025-06-30
+
+- Перенесена конфигурация Next.js с `next.config.ts` на `next.config.js` для совместимости с Docker и Next.js 15.
