+++ STATE.md
@@ -19,7 +19,7 @@
 - [x] Очистка тестов, убраны неиспользуемые импорты, flake8 проходит ✅
 - [x] Единый стиль оформления кода: black, isort, flake8 ✅
 - [x] Pre-commit настроен, форматирование кода стабильно ✅
-- [x] Проведён аудит next.config.ts, устаревший next.config.js отсутствует
+- [x] next.config.ts заменён на next.config.js для совместимости
 - [x] Удалены тестовые роуты frontend/src/app/test
 - [x] Создан CONTRIBUTING.md с инструкциями для разработчиков
 - [x] Обновлён CHANGELOG.md с последними изменениями
