+++ docs/migration-summary.md
@@ -1,0 +1,15 @@
+# Сводка миграции next.config
+
+## Изменения
+- Создан `frontend/next.config.js` и удалён `next.config.ts`.
+- Обновлены README и комментарии, все ссылки указывают на новый файл.
+- Настройки полностью сохранены.
+- Документация дополнена файлом `docs/next-config-migration.md`.
+
+## Тесты
+- `npm test` и `pytest` выполнены успешно после изменений.
+- Сборка `next build` проходит без ошибок.
+- `docker compose up --build` поднимает фронтенд без проблем.
+
+## Как откатить
+Если потребуется вернуть TypeScript вариант, восстановите `frontend/next.config.ts` из истории git и удалите `next.config.js`. Не забудьте вернуть упоминания в документации и конфигурациях.
