+++ docs/next-config-migration.md
@@ -1,0 +1,20 @@
+# Миграция конфигурации Next.js
+
+Проект использовал файл `next.config.ts`. В среде Docker при использовании Next.js 15 этот формат перестал поддерживаться напрямую. Поэтому конфигурация была переписана на JavaScript.
+
+## Почему был удалён `next.config.ts`
+- Next.js 15 не поддерживает загрузку TypeScript-конфига при запуске через `node` внутри контейнера.
+- Docker-образы становятся проще: не требуется трансформация TypeScript на этапе запуска.
+
+## Как работает `next.config.js`
+Файл `frontend/next.config.js` экспортирует объект конфигурации через `module.exports`. Все настройки эквивалентны предыдущему варианту:
+- `rewrites` проксирует запросы `/api/v1/*` и `/media/*` на адрес бэкенда из `BACKEND_URL`.
+- `images.remotePatterns` разрешает загрузку изображений с того же хоста, что и API.
+- Включён `reactStrictMode`.
+- В секции `typescript` сохранена опция `ignoreBuildErrors`.
+
+## Как изменять конфиг
+Редактируйте `next.config.js` как обычный JS-файл. Если требуется добавить новые параметры, просто расширяйте экспортируемый объект. Для подсказок редактора сохранены JSDoc комментарии с типом `NextConfig`.
+
+## Использование в Docker и CI/CD
+При сборке Docker-образа переменные окружения (например, `BACKEND_URL`) передаются в контейнер через `docker-compose.yml` или аргументы `docker build`. В CI конфиг автоматически подхватывается при `npm run build` без дополнительной компиляции.
